* cfgloop.c (flow_loop_dump): Cast nit to uint64_t and print it using 	PRIu64 instead of lu.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@237772 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/cfgloop.c
@@ -141,7 +141,8 @@ flow_loop_dump (const struct loop *loop, FILE *file,
       bool read_profile_p;
       gcov_type nit = expected_loop_iterations_unbounded (loop, &read_profile_p);
       if (read_profile_p && !loop->any_estimate)
-	fprintf (file, ";;  profile-based iteration count: %lu\n", nit);
+	fprintf (file, ";;  profile-based iteration count: %" PRIu64 "\n",
+		 (uint64_t) nit);
     }
 
   fprintf (file, ";;  nodes:");
